feat: add admin session

+++ php/participant/participant_delete_participant.php
@@ -13,22 +13,59 @@
 <body>
     <?php
     include "../connect.php";
-    // variables
-    $participantid = $_GET["participantid"];
-    
 
-    // sql
-    $sql = "DELETE FROM participants WHERE participantid = '$participantid'";
-    $result = mysqli_query($con, $sql) or die("Error deleting data due to " . mysqli_error($con));
+    session_start();
+    if (isset($_SESSION["username"])) :
+        $userid = $_SESSION["userid"];
+        $sessionusername = $_SESSION["username"];
+        $role = $_SESSION["role"];
 
-    // response
-    if ($result) {
-        echo "You have successfully deleted the participant.";
-        header("location:participant.php");
-    } else {
-        echo "Failed to delete the participant.";
-        header("location:participant.php");
-    }
+        if ($role == "admin") :
+
+            // from participant.php
+            $participantid = $_GET["participantid"];
+
+            // delete participant from participants table
+            $sql = "DELETE FROM participants WHERE participantid = '$participantid'";
+            $result = mysqli_query($con, $sql) or die("Error deleting data due to " . mysqli_error($con));
+
+            // response
+            if ($result) {
+                echo "You have successfully deleted the participant.";
+                header('Location: ' . $_SERVER['HTTP_REFERER']);
+            } else {
+                echo "Failed to delete the participant.";
+                header('Location: ' . $_SERVER['HTTP_REFERER']);
+            }
+        else : ?>
+
+            <div class="container">
+                <main>
+                    <div class="py-5 text-center">
+                        <h2>Error</h2>
+                        <p>Restricted to admin. Redirecting you to homepage...</p>
+                    </div>
+                </main>
+            </div>
+
+        <?php
+            header("refresh:3;url='../home/home.php'");
+            exit;
+        endif;
+    else :
+        ?>
+
+        <div class="container">
+            <main>
+                <div class="py-5 text-center">
+                    <h2>Error</h2>
+                    <p>Unauthorized access.</p>
+                    <a class="btn btn-primary" href="../signin/signin.php">Sign In</a>
+                </div>
+            </main>
+        </div>
+
+    <?php endif;
     ?>
 
     <script src="../../js/colormodes.js"></script>
